Trying to make samle and boxyhq live together
