+++ logicle/app/layout.tsx
@@ -47,6 +47,7 @@ export default async function RootLayout({
     maxImgAttachmentDimPx: env.chat.attachments.maxImgDimPx,
     enableApiKeys: env.apiKeys.enable,
     appUrl: env.appUrl,
+    useSaml2Js: env.saml.useSaml2,
   }
 
   return (
