+++ logicle/lib/auth/InvalidCredentialError.ts
@@ -1,0 +1,8 @@
+import { CredentialsSignin } from 'next-auth'
+
+export class InvalidCredentialsError extends CredentialsSignin {
+  constructor(code: string) {
+    super(code)
+    this.code = code
+  }
+}
